Added functionality for "shared" checkbox
There is now a "shared" and an "element" prop on checkboxes which allows them to reference the same filter (ie. same "name" and "value") for array-type filters (like the "price" filter I also added). Set the same name for all shared checkboxes, and specify which element of the filter array each controls, and viola!